@@ -13,29 +13,29 @@
     "@react-native-community/blur": "3.6.0",
     "@react-native-community/masked-view": "0.1.10",
     "@react-native-community/netinfo": "5.9.7",
-    "@react-navigation/native": "5.7.6",
-    "@react-navigation/stack": "5.9.3",
+    "@react-navigation/native": "5.8.6",
+    "@react-navigation/stack": "5.12.3",
     "react": "16.13.1",
     "react-native": "0.63.3",
-    "react-native-document-picker": "4.0.0",
+    "react-native-document-picker": "4.1.0",
     "react-native-gesture-handler": "1.8.0",
     "react-native-get-random-values": "1.5.0",
     "react-native-image-crop-picker": "0.35.1",
-    "react-native-reanimated": "2.0.0-alpha.7",
+    "react-native-reanimated": "2.0.0-alpha.8",
     "react-native-safe-area-context": "3.1.8",
-    "react-native-screens": "2.11.0",
+    "react-native-screens": "2.13.0",
     "react-native-svg": "12.1.0",
-    "stream-chat": "2.7.0",
+    "stream-chat": "2.7.3",
     "stream-chat-react-native": "link:../../native-package"
   },
   "devDependencies": {
     "@babel/core": "7.12.3",
-    "@babel/runtime": "7.12.1",
+    "@babel/runtime": "7.12.5",
     "@react-native-community/eslint-config": "2.0.0",
-    "babel-jest": "26.6.1",
-    "eslint": "7.11.0",
-    "jest": "26.6.1",
-    "metro-react-native-babel-preset": "0.63.0",
+    "babel-jest": "26.6.3",
+    "eslint": "7.12.1",
+    "jest": "26.6.3",
+    "metro-react-native-babel-preset": "0.64.0",
     "react-test-renderer": "16.13.1"
   },
   "jest": {
