@@ -77,7 +77,10 @@ registerNativeHandlers({
     try {
       let res = await ImagePicker.openPicker({
         compressImageQuality,
+        forceJpg: true,
+        includeBase64: Platform.OS === 'ios',
         maxFiles: maxNumberOfFiles || undefined,
+        mediaType: 'photo',
         multiple: true,
         writeTempFile: false,
       });
@@ -90,7 +93,10 @@ registerNativeHandlers({
       return {
         cancelled: false,
         images: res.map((image) => ({
-          uri: Platform.OS === 'ios' ? image.sourceURL : image.path,
+          uri:
+            Platform.OS === 'ios'
+              ? image.sourceURL || `data:${image.mime};base64,${image.data}`
+              : image.path,
         })),
       };
     } catch (err) {
