@@ -56,7 +56,7 @@ export const LoadingErrorIndicator: React.FC<LoadingErrorProps> = (props) => {
       return (
         <LoadingErrorWrapper
           onPress={retry}
-          text={t('Error loading channel list ...')}
+          text={t('Error loading channel list...')}
         >
           <RetryText>⟳</RetryText>
         </LoadingErrorWrapper>
@@ -64,7 +64,7 @@ export const LoadingErrorIndicator: React.FC<LoadingErrorProps> = (props) => {
     case 'message':
       return (
         <LoadingErrorWrapper
-          text={t('Error loading messages for this channel ...')}
+          text={t('Error loading messages for this channel...')}
         />
       );
     default:
