@@ -12,6 +12,7 @@ import {
   Attachment,
   logChatPromiseExecution,
   SendFileAPIResponse,
+  StreamChat,
   Message as StreamMessage,
   UserResponse,
 } from 'stream-chat';
@@ -770,7 +771,9 @@ export const MessageInput = <
         attachments,
         mentioned_users: mentionedUsers,
         text: prevText,
-      } as StreamMessage<At, Me, Us>;
+      } as Parameters<
+        StreamChat<At, Ch, Co, Ev, Me, Re, Us>['updateMessage']
+      >[0];
 
       // TODO: Remove this line and show an error when submit fails
       clearEditingState();
@@ -824,7 +827,7 @@ export const MessageInput = <
         await client.updateMessage({
           ...editing,
           text,
-        } as StreamMessage<At, Me, Us>);
+        } as Parameters<StreamChat<At, Ch, Co, Ev, Me, Re, Us>['updateMessage']>[0]);
       }
 
       setText('');
