@@ -65,9 +65,7 @@ export type MessagesContextValue<
 > = {
   clearEditingState: () => void;
   editing: boolean | Message<At, Ch, Co, Ev, Me, Re, Us>;
-  editMessage: (
-    updatedMessage: StreamMessage<At, Me, Us>,
-  ) => ReturnType<StreamChat<At, Ch, Co, Ev, Me, Re, Us>['updateMessage']>;
+  editMessage: StreamChat<At, Ch, Co, Ev, Me, Re, Us>['updateMessage'];
   emojiData: Array<{
     icon: string;
     id: string;
