@@ -145,8 +145,9 @@ export class KeyboardCompatibleView extends React.Component<
   };
 
   unsetKeyboardListeners = () => {
-    this._subscriptions.forEach((subscription) => {
+    this._subscriptions = this._subscriptions.filter((subscription) => {
       subscription.remove();
+      return false;
     });
   };
 
