@@ -5,7 +5,6 @@ import {
 } from '../utils/insertDates';
 
 import type { ImmutableDate } from 'seamless-immutable';
-import type { UserResponse } from 'stream-chat';
 
 import type { ChannelContextValue } from '../../../contexts/channelContext/ChannelContext';
 import type {
@@ -28,15 +27,16 @@ export const getReadStates = <
   Re extends UnknownType = DefaultReactionType,
   Us extends UnknownType = DefaultUserType
 >(
+  clientUserId: string | undefined,
   messages: InsertDatesResponse<At, Ch, Co, Ev, Me, Re, Us>,
   read?: ChannelContextValue<At, Ch, Co, Ev, Me, Re, Us>['read'],
 ) => {
   const readData = messages.reduce((acc, cur) => {
     if (!isDateSeparator(cur) && cur.id) {
-      acc[cur.id] = [];
+      acc[cur.id] = false;
     }
     return acc;
-  }, {} as { [key: string]: UserResponse<Us>[] });
+  }, {} as { [key: string]: boolean | number });
 
   const filteredMessagesReversed = messages
     .filter((msg) => !isDateSeparator(msg) && msg.updated_at)
@@ -48,36 +48,60 @@ export const getReadStates = <
 
   if (read) {
     /**
-     * Channel read state is stored by user
+     * Channel read state is stored by user and we only care about users who aren't the client
      */
-    for (const readState of Object.values(read)) {
-      /**
-       * If no last read continue
-       */
-      if (!readState.last_read) {
-        continue;
-      }
+    const readList = read?.asMutable ? read.asMutable() : read;
+    if (clientUserId) {
+      delete readList[clientUserId];
+    }
+    const members = Object.values(readList);
 
-      /**
-       * Array is in reverse order so newest message is at 0,
-       * we find the index of the first message that is older
-       * than the last read and then set last read to that, or
-       * if there are no newer messages, the first message is
-       * last read message.
-       */
-      const userLastReadMsgId = filteredMessagesReversed.find(
-        (msg) => msg.updated_at < readState.last_read,
-      )?.id;
+    /**
+     * Array is in reverse order so newest message is at 0,
+     * we find the index of the first message that is older
+     * than the last read and then set last read to that, or
+     * if there are no newer messages, the first message is
+     * last read message.
+     */
+    for (const message of filteredMessagesReversed) {
+      if (!members.length) {
+        readData[message.id] = true;
+      } else {
+        for (const member of members) {
+          /**
+           * If no last read continue
+           */
+          if (!member.last_read) {
+            continue;
+          }
 
-      /**
-       * If there there is a last read message add the user
-       * to the array of last reads for that message
-       */
-      if (userLastReadMsgId) {
-        readData[userLastReadMsgId] = [
-          ...readData[userLastReadMsgId],
-          readState.user as UserResponse<Us>,
-        ];
+          /**
+           * If there there is a last read message add the user
+           * to the array of last reads for that message and remove
+           * the user from the list of users being checked
+           */
+          if (message.updated_at < member.last_read) {
+            const currentMessageReadData = readData[message.id];
+            /**
+             * if this is a direct message the length will be 1
+             * as we already deleted the current user from the object
+             */
+            if (Object.keys(readList).length === 1) {
+              readData[message.id] = true;
+            } else {
+              readData[message.id] =
+                typeof currentMessageReadData === 'boolean'
+                  ? 1
+                  : currentMessageReadData + 1;
+            }
+            const userIndex = members.findIndex(
+              ({ user }) => user.id === message.user?.id,
+            );
+            if (userIndex !== -1) {
+              members.splice(userIndex, 1);
+            }
+          }
+        }
       }
     }
   }
