@@ -1,7 +1,8 @@
 export * from './channelContext/ChannelContext';
 export * from './chatContext/ChatContext';
 export * from './keyboardContext/KeyboardContext';
-export * from './messageContentContext/MessageContentContext';
+export * from './messageContext/MessageContext';
+export * from './messageOverlayContext/MessageOverlayContext';
 export * from './messagesContext/MessagesContext';
 export * from './overlayContext/OverlayContext';
 export * from './suggestionsContext/SuggestionsContext';
