@@ -1,5 +1,9 @@
 import React from 'react';
+import Animated, { useAnimatedProps } from 'react-native-reanimated';
 import Svg, { Path, PathProps, SvgProps } from 'react-native-svg';
+
+const AnimatedPath = Animated.createAnimatedComponent(Path);
+const AnimatedSvg = Animated.createAnimatedComponent(Svg);
 
 export type IconProps = Partial<SvgProps> &
   Omit<RootPathProps, 'd'> & {
@@ -9,17 +13,25 @@ export type IconProps = Partial<SvgProps> &
   };
 
 export const RootSvg: React.FC<IconProps> = (props) => {
-  const { backgroundFill = 'none', children, height = 24, width = 24 } = props;
+  const {
+    backgroundFill = 'none',
+    children,
+    height = 24,
+    onPress,
+    viewBox = '0 0 24 24',
+    width = 24,
+  } = props;
+  const animatedProps = useAnimatedProps(() => ({
+    fill: backgroundFill,
+    height,
+    onPress,
+    viewBox,
+    width,
+  }));
   return (
-    <Svg
-      fill={backgroundFill}
-      height={height}
-      viewBox='0 0 24 24'
-      width={width}
-      {...props}
-    >
+    <AnimatedSvg animatedProps={animatedProps} {...props}>
       {children}
-    </Svg>
+    </AnimatedSvg>
   );
 };
 
@@ -31,13 +43,12 @@ export type RootPathProps = {
 
 export const RootPath: React.FC<RootPathProps> = (props) => {
   const { d, pathFill = '#000000', pathOpacity } = props;
-  return (
-    <Path
-      clipRule='evenodd'
-      d={d}
-      fill={pathFill}
-      fillRule='evenodd'
-      opacity={pathOpacity}
-    />
-  );
+  const animatedProps = useAnimatedProps(() => ({
+    clipRule: 'evenodd',
+    d,
+    fill: pathFill,
+    fillRule: 'evenodd',
+    opacity: pathOpacity,
+  }));
+  return <AnimatedPath animatedProps={animatedProps} />;
 };
